feat!: bypass payment, now just add to library
