+++ src/app/(with-nav)/marketplace/components/itemcard.tsx
@@ -2,7 +2,6 @@ import Image from "next/image";
 import { LuCircleCheckBig, LuDownload, LuPlus } from "react-icons/lu";
 
 import { Button } from "@/ui/button";
-import { addToCart } from "@/app/(with-nav)/marketplace/cart.actions";
 import { Stars } from "./stars";
 import { CardBadge } from "./card-badge";
 import Link from "next/link";
@@ -10,6 +9,7 @@ import { startTransition, useActionState } from "react";
 import { deletePublication } from "../../profile/actions/deletePost";
 import { Dialog, DialogClose } from "@/ui/dialog";
 import { LuTrash2 } from "react-icons/lu";
+import { addToLibrary } from "../actions";
 
 type ItemCardProps = {
   itemType: "Shader" | "Project"; // TODO group in the future
@@ -17,7 +17,7 @@ type ItemCardProps = {
   title: string;
   price: number;
   downloads: number;
-  inCart: boolean;
+  isOwned: boolean;
   username?: string;
   isOwn?: boolean;
   category?: string;
@@ -34,7 +34,7 @@ export default function ItemCard({
   title,
   price,
   downloads,
-  inCart,
+  isOwned,
   username,
   isOwn,
   category,
@@ -48,13 +48,11 @@ export default function ItemCard({
     Date.now() - new Date(createdAt).getTime() < 7 * 24 * 60 * 60 * 1000;
 
   // TODO pending behaviour
-  const [addToCartState, addToCartAction, addToCartPending] = useActionState(
-    async (_prev: null, formData: FormData) => {
-      await addToCart(formData);
+  const [addToLibraryState, addToLibraryAction, addToLibraryPending] =
+    useActionState(async (_prev: null, formData: FormData) => {
+      await addToLibrary(formData);
       return null;
-    },
-    null,
-  );
+    }, null);
 
   return (
     <div className="glass glass-border p-4 rounded-2xl relative hover:bg-current/1">
@@ -138,7 +136,7 @@ export default function ItemCard({
               </DialogClose>
             </form>
           </Dialog>
-        ) : inCart ? (
+        ) : isOwned ? (
           <div className="flex justify-center items-center h-13.5 text-base/5 font-semibold rounded-lg border border-current/15 select-none">
             <LuCircleCheckBig className="inline mr-2 text-emerald-600" />
             In library
@@ -146,7 +144,7 @@ export default function ItemCard({
         ) : (
           <form
             action={(formData) =>
-              startTransition(() => addToCartAction(formData))
+              startTransition(() => addToLibraryAction(formData))
             }
           >
             <input type="hidden" name="itemId" value={id} />
